Add viewpager2 and move patient list logic to patient list fragment
