+++ app/src/main/java/org/smartregister/fhircore/fragment/PatientListFragment.kt
@@ -1,0 +1,76 @@
+package org.smartregister.fhircore.fragment
+
+import android.content.Intent
+import android.os.Bundle
+import android.util.Log
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import android.widget.TextView
+import android.widget.Toast
+import androidx.core.view.GravityCompat
+import androidx.drawerlayout.widget.DrawerLayout
+import androidx.lifecycle.ViewModelProvider
+import androidx.recyclerview.widget.RecyclerView
+import com.google.android.fhir.FhirEngine
+import org.smartregister.fhircore.FhirApplication
+import org.smartregister.fhircore.PatientListViewModel
+import org.smartregister.fhircore.PatientListViewModelFactory
+import org.smartregister.fhircore.R
+import org.smartregister.fhircore.activity.PatientDetailActivity
+import org.smartregister.fhircore.adapter.PatientItemRecyclerViewAdapter
+
+class PatientListFragment : Fragment() {
+
+    private lateinit var patientListViewModel: PatientListViewModel
+    private lateinit var fhirEngine: FhirEngine
+
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
+                              savedInstanceState: Bundle?): View? {
+        return inflater.inflate(R.layout.fragment_patient_list, container, false)
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        fhirEngine = FhirApplication.fhirEngine(requireContext())
+
+        val recyclerView = view.findViewById<RecyclerView>(R.id.patient_list)
+        val adapter = PatientItemRecyclerViewAdapter(this::onPatientItemClicked)
+        recyclerView.adapter = adapter
+
+        requireActivity().findViewById<TextView>(R.id.tv_sync).setOnClickListener {
+            requireActivity().findViewById<DrawerLayout>(R.id.drawer_layout).closeDrawer(GravityCompat.START)
+            syncResources()
+        }
+
+        patientListViewModel =
+                ViewModelProvider(this, PatientListViewModelFactory(requireActivity().application, fhirEngine))
+                        .get(PatientListViewModel::class.java)
+
+        patientListViewModel
+                .getSearchedPatients()
+                .observe(
+                        viewLifecycleOwner,
+                        {
+                            Log.d("PatientListActivity", "Submitting ${it.count()} patient records")
+                            adapter.submitList(it)
+                        }
+                )
+
+        super.onViewCreated(view, savedInstanceState)
+    }
+
+    // Click handler to help display the details about the patients from the list.
+    private fun onPatientItemClicked(patientItem: PatientListViewModel.PatientItem) {
+        val intent =
+                Intent(requireActivity(), PatientDetailActivity::class.java).apply {
+                    putExtra(PatientDetailFragment.ARG_ITEM_ID, patientItem.id)
+                }
+        this.startActivity(intent)
+    }
+
+    private fun syncResources() {
+        Toast.makeText(requireContext(), "Syncing...", Toast.LENGTH_LONG).show()
+        patientListViewModel.searchPatients()
+    }
+}
