+++ app/src/main/res/layout/activity_patient_list.xml
@@ -47,13 +47,12 @@
 
         </androidx.appcompat.widget.Toolbar>
 
-        <include
-            android:id="@+id/patient_list"
-            layout="@layout/patient_list"
+        <androidx.viewpager2.widget.ViewPager2
+            android:id="@+id/patient_list_pager"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             app:layout_constraintTop_toBottomOf="@id/patient_register_toolbar"
-            tools:layout_editor_absoluteX="0dp" />
+            app:layout_constraintBottom_toTopOf="@id/btn_register_new_patient" />
 
         <Button
             android:id="@+id/btn_register_new_patient"
