+++ app/src/main/res/values/strings.xml
@@ -28,4 +28,6 @@
 
     <string name="submit_button_text">Submit</string>
     <string name="search_hint">Search name or ID</string>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
